feat(web): Update diagram outline UI
