+++ app/web/src/organisms/ComponentTree.vue
@@ -4,6 +4,7 @@
       v-for="(node, index) in treeData"
       :key="index"
       :component="node"
+      :class="nodeClass ?? ''"
       @select="(componentId) => emit('select', componentId)"
       @multiselect="(componentId) => emit('multiselect', componentId)"
       @pan="(componentId) => emit('pan', componentId)"
@@ -13,9 +14,12 @@
 
 <script lang="ts" setup>
 import { ComponentTreeNode } from "@/store/components.store";
-import TreeNode from "@/organisms/TreeNode.vue";
+import TreeNode from "@/organisms/ComponentTreeNode.vue";
 
-const props = defineProps<{ treeData: ComponentTreeNode[] }>();
+const props = defineProps<{
+  treeData: ComponentTreeNode[];
+  nodeClass?: string;
+}>();
 
 const emit = defineEmits<{
   (e: "select", componentId: string): void;
