--- app/web/src/organisms/TreeNode.vue
@@ -1,176 +1,0 @@
-<template>
-  <li
-    class="border-b-2 dark:border-neutral-600 cursor-pointer"
-    @click.exact.stop="emit('select', component.id)"
-    @click.meta.stop="emit('multiselect', component.id)"
-    @click.ctrl.stop="emit('multiselect', component.id)"
-    @dblclick.stop="emit('pan', component.id)"
-  >
-    <template v-if="component.children">
-      <Disclosure v-if="component.matchesFilter" v-slot="{ open }" default-open>
-        <div
-          class="flex flex-row"
-          :class="
-            isSelected
-              ? ['bg-action-500 text-white']
-              : ['hover:bg-action-400 hover:text-white']
-          "
-        >
-          <DisclosureButton>
-            <Icon
-              :name="open ? 'chevron--down' : 'chevron--right'"
-              size="sm"
-              class="dark:text-white flex-shrink-0 block"
-            />
-          </DisclosureButton>
-          <span
-            :class="
-              isSelected
-                ? ['bg-action-500 text-white']
-                : ['hover:bg-action-400 hover:text-white']
-            "
-            :style="{
-              'border-color': component.color || colors.neutral[400],
-            }"
-            class="px-2 py-2 border-l-8 group flex flex-col items-baseline w-full"
-          >
-            <span
-              class="whitespace-nowrap text-ellipsis overflow-hidden shrink leading-tight"
-              >{{ component.displayName || "si-123" }}</span
-            >
-            <i
-              :class="
-                isSelected
-                  ? ['bg-action-500 text-white']
-                  : ['text-neutral-500 group-hover:text-white']
-              "
-              class="text-sm pl-1 flex-none"
-            >
-              Frame
-            </i>
-          </span>
-          <div class="flex flex-row items-center">
-            <Icon
-              v-if="statusIcons.change"
-              :name="statusIcons.change.icon"
-              :tone="statusIcons.change.tone"
-            />
-            <div v-else class="w-6 h-6" />
-            <Icon
-              v-if="statusIcons.qualification"
-              :name="statusIcons.qualification.icon"
-              :tone="statusIcons.qualification.tone"
-            />
-            <div v-else class="w-6 h-6" />
-            <Icon
-              v-if="statusIcons.confirmation"
-              :name="statusIcons.confirmation.icon"
-              :tone="statusIcons.confirmation.tone"
-            />
-            <div v-else class="w-6 h-6" />
-          </div>
-        </div>
-        <DisclosurePanel>
-          <Tree
-            :tree-data="component.children"
-            class="pl-8"
-            @select="(componentId) => emit('select', componentId)"
-            @multiselect="(componentId) => emit('multiselect', componentId)"
-            @pan="(componentId) => emit('pan', componentId)"
-          />
-        </DisclosurePanel>
-      </Disclosure>
-      <template v-else>
-        <Tree
-          :tree-data="component.children"
-          class="pl-8"
-          @select="(componentId) => emit('select', componentId)"
-          @multiselect="(componentId) => emit('multiselect', componentId)"
-          @pan="(componentId) => emit('pan', componentId)"
-        />
-      </template>
-    </template>
-    <div
-      v-else-if="component.matchesFilter"
-      class="flex flex-row items-center"
-      :class="
-        isSelected
-          ? ['bg-action-500 text-white']
-          : ['hover:bg-action-400 hover:text-white']
-      "
-    >
-      <span
-        :style="{
-          'border-color': component.color || colors.neutral[400],
-        }"
-        class="w-full px-2 py-2 border-l-8 group flex flex-col items-baseline ml-5"
-      >
-        <span
-          class="whitespace-nowrap text-ellipsis overflow-hidden shrink leading-tight"
-          >{{ component.displayName || "si-123" }}</span
-        >
-        <i
-          :class="
-            isSelected
-              ? ['bg-action-500 text-white']
-              : ['text-neutral-500 group-hover:text-white']
-          "
-          class="text-sm pl-1 flex-none"
-        >
-          {{ component.schemaName }}
-        </i>
-      </span>
-      <div class="flex flex-row items-center">
-        <Icon
-          v-if="statusIcons.change"
-          :name="statusIcons.change.icon"
-          :tone="statusIcons.change.tone"
-        />
-        <div v-else class="w-6 h-6" />
-        <Icon
-          v-if="statusIcons.qualification"
-          :name="statusIcons.qualification.icon"
-          :tone="statusIcons.qualification.tone"
-        />
-        <div v-else class="w-6 h-6" />
-        <Icon
-          v-if="statusIcons.confirmation"
-          :name="statusIcons.confirmation.icon"
-          :tone="statusIcons.confirmation.tone"
-        />
-        <div v-else class="w-6 h-6" />
-      </div>
-    </div>
-  </li>
-</template>
-
-<script lang="ts" setup>
-import { computed } from "vue";
-import { Disclosure, DisclosurePanel, DisclosureButton } from "@headlessui/vue";
-import clsx from "clsx";
-import _ from "lodash";
-import {
-  ComponentTreeNode,
-  useComponentsStore,
-} from "@/store/components.store";
-import { colors } from "@/utils/design_token_values";
-import Icon from "@/ui-lib/icons/Icon.vue";
-import Tree from "@/organisms/Tree.vue";
-
-const props = defineProps<{ component: ComponentTreeNode }>();
-
-const emit = defineEmits<{
-  (e: "select", componentId: string): void;
-  (e: "multiselect", componentId: string): void;
-  (e: "pan", componentId: string): void;
-}>();
-
-const componentsStore = useComponentsStore();
-
-const selectedComponentId = computed(() => componentsStore.selectedComponentId);
-const isSelected = computed(() => {
-  return _.includes(componentsStore.selectedComponentIds, props.component.id);
-});
-
-const statusIcons = computed(() => props.component.statusIcons ?? {});
-</script>
